backend/requirements.txt
django
django-cors-headers
PyJWT
dj-database-url==3.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
whitenoise==6.11.0
